RHEL 7: v2v: Select correct qemu binary for -o qemu mode (RHBZ#1147313).
RHEL 7 does not have qemu-system-x86_64 (etc), and in addition the
qemu binary is located in /usr/libexec.  Encode the path to this
binary directly in the script.

Note that we don't support people running qemu directly like this.
It's just for quick testing of converted VMs, and to help us with
support cases.